update tests
Divide requirements files for tests and static code analysis.
Split tests into:
- software tests
- system tests (to be created)

Conftest files optimized.
