+++ .travis.yml
@@ -7,11 +7,11 @@ python:
 
 install:
   - "pip install -r requirements.txt"
-  - "pip install -r tests/requirements_for_tests.txt"
+  - "pip install -r tests/requirements_for_software_tests.txt"
   - "pip install coveralls"
 
 script:
-  - "pytest --cov-report term-missing --cov=uds tests"
-  - "coverage run --source=uds -m pytest tests/"
+  - "pytest --cov-report term-missing --cov=uds tests/software_tests"
+  - "coverage run --source=uds -m pytest tests/software_tests/"
  
 after_sucess: coveralls
