+++ tests/requirements_for_software_tests.txt
@@ -1,0 +1,3 @@
+pytest>=6.2.2
+pytest-cov>=2.11.1
+mock>=4.0.2
