+++ tests/requirements_for_static_code_analysis.txt
@@ -1,0 +1,5 @@
+prospector==1.3.1
+mypy==0.800
+vulture==2.3
+pyroma==2.6
+bandit==1.7.0
