--- tests/requirements_for_tests.txt
@@ -1,7 +1,0 @@
-pytest>=6.2.2
-pytest-cov>=2.11.1
-mock>=4.0.2
-prospector==1.3.1
-mypy==0.800
-vulture==2.3
-pyroma==2.6
