+++ tests/software_tests/__init__.py
@@ -1,0 +1,11 @@
+"""
+Software tests for the package.
+
+Software tests includes:
+- unit tests (SWE.4 [1])
+- integration tests (SWE.5 [1])
+- software qualification tests (SWE.6 [1])
+
+References:
+- [1] - Automotive SPICE 3.0 http://www.automotivespice.com/fileadmin/software-download/Automotive_SPICE_PAM_30.pdf
+"""
