+++ tests/software_tests/messages/test_base_message.py
