+++ tests/software_tests/server/test_response_rule.py
